Restructuring Card view component
