+++ lib/core/card-view/card-view.module.scss
@@ -1,0 +1,9 @@
+@import './components/card-view-dateitem/card-view-dateitem.component';
+@import './components/card-view-textitem/card-view-textitem.component';
+@import './components/card-view/card-view.component';
+
+@mixin adf-card-view-module-theme($theme) {
+    @include adf-card-view-dateitem-theme($theme);
+    @include adf-card-view-textitem-theme($theme);
+    @include adf-card-view-theme($theme);
+}
