+++ lib/core/card-view/components/card-view-dateitem/card-view-dateitem.component.spec.ts
@@ -21,10 +21,10 @@ import { MatDatepickerModule, MatInputModule, MatNativeDateModule } from '@angul
 import { By } from '@angular/platform-browser';
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 import moment from 'moment-es6';
-import { AppConfigService } from '../index';
-import { CardViewDateItemModel } from '../models/card-view-dateitem.model';
-import { CardViewUpdateService } from '../services/card-view-update.service';
-import { TranslateLoaderService } from '../services/translate-loader.service';
+import { AppConfigService } from '../../../index';
+import { CardViewDateItemModel } from '../../models/card-view-dateitem.model';
+import { CardViewUpdateService } from '../../services/card-view-update.service';
+import { TranslateLoaderService } from '../../../services/translate-loader.service';
 
 import { CardViewDateItemComponent } from './card-view-dateitem.component';
