+++ lib/core/card-view/components/card-view-item-dispatcher/card-view-item-dispatcher.component.spec.ts
@@ -21,10 +21,10 @@ import { Component, Input, SimpleChange } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { BrowserDynamicTestingModule } from '@angular/platform-browser-dynamic/testing';
-import { CardViewItem } from '../interface/card-view-item.interface';
-import { CardItemTypeService } from '../services/card-item-types.service';
-import { CardViewContentProxyDirective } from './card-view-content-proxy.directive';
-import { CardViewItemDispatcherComponent } from './card-view-item-dispatcher.component';
+import { CardViewItem } from '../../interfaces/card-view-item.interface';
+import { CardItemTypeService } from '../../services/card-item-types.service';
+import { CardViewContentProxyDirective } from '../../directives/card-view-content-proxy.directive';
+import { CardViewItemDispatcherComponent } from '../card-view-item-dispatcher/card-view-item-dispatcher.component';
 
 @Component({
     selector: 'whatever-you-want-to-have',
