+++ lib/core/card-view/components/card-view-mapitem/card-view-mapitem.component.html
