+++ lib/core/card-view/components/card-view-mapitem/card-view-mapitem.component.ts
@@ -16,17 +16,26 @@
  */
 
 import { Component, Input } from '@angular/core';
-import { CardViewItem } from '../interface/card-view-item.interface';
+import { CardViewMapItemModel } from '../../models/card-view-mapitem.model';
+import { CardViewUpdateService } from '../../services/card-view-update.service';
 
 @Component({
-    selector: 'adf-card-view',
-    templateUrl: './card-view.component.html',
-    styleUrls: ['./card-view.component.scss']
+    selector: 'adf-card-view-mapitem',
+    templateUrl: './card-view-mapitem.component.html',
+    styleUrls: ['./card-view-mapitem.component.scss']
 })
-export class CardViewComponent {
-    @Input()
-    properties: CardViewItem [];
 
+export class CardViewMapItemComponent {
     @Input()
-    editable: boolean;
+    property: CardViewMapItemModel;
+
+    constructor(private cardViewUpdateService: CardViewUpdateService) {}
+
+    isClickable() {
+        return this.property.clickable;
+    }
+
+    clicked(): void {
+        this.cardViewUpdateService.clicked(this.property);
+    }
 }
