+++ lib/core/card-view/components/card-view-textitem/card-view-textitem.component.html
