+++ lib/core/card-view/components/card-view/card-view.component.html
@@ -5,4 +5,3 @@
         </div>
     </div>
 </div>
-
