+++ lib/core/card-view/components/card-view/card-view.component.scss
