+++ lib/core/card-view/directives/card-view-content-proxy.directive.ts
