+++ lib/core/card-view/interfaces/card-view-item.interface.ts
