+++ lib/core/card-view/models/card-view-baseitem.model.ts
