+++ lib/core/card-view/models/card-view-textitem.model.spec.ts
