+++ lib/core/card-view/services/card-view-update.service.ts
