+++ lib/core/interface/public-api.ts
@@ -16,5 +16,4 @@
  */
 
 export * from './authentication.interface';
-export * from './card-view-item.interface';
 export * from './injection.tokens';
