+++ lib/core/styles/_index.scss
@@ -2,10 +2,7 @@
 @import './default-class';
 @import './theming';
 
-
-@import '../card-view/card-view-dateitem.component';
-@import '../card-view/card-view-textitem.component';
-@import '../card-view/card-view.component';
+@import '../card-view/card-view.module';
 @import '../collapsable/accordion-group.component';
 @import '../datatable/components/datatable/datatable.component';
 @import '../form/components/widgets/container/container.widget';
@@ -25,9 +22,7 @@
 @mixin adf-core-theme($theme) {
     @include adf-colors-theme($theme);
     @include adf-default-class-theme($theme);
-    @include adf-card-view-dateitem-theme($theme);
-    @include adf-card-view-textitem-theme($theme);
-    @include adf-card-view-theme($theme);
+    @include adf-card-view-module-theme($theme);
     @include adf-accordion-theme($theme);
     @include adf-datatable-theme($theme);
     @include adf-form-container-widget-theme($theme);
